Food Recognition and Recipe Suggestion/Backend/requirements.txt
Flask==2.0.2
groq==0.14.0
pathvalidate==3.2.3
pillow==11.1.0
pydantic==2.10.5
jsonschema==4.23.0
requests==2.31.0
python-dotenv==1.2.1